+++ crates/domain/src/model/mod.rs
@@ -1,3 +1,4 @@
+pub mod api_error;
 pub mod cookie;
 pub mod download;
 pub mod music_info;